Proof of concept.
